stm_mcu: Fix "ref without flash", remove STM8
Some STM8 MCUs don't follow the naming scheme of the STM32 MCUs, so the
grouping into categories was wrong. I tried to find a "system" for a
long time, but it would be a huge amount of work to get it fixed
properly.

The simplest way for now is to exclude STM8 MCUs from being generated.