so: Bump version of TSSOP packages (#102)
The TSSOP packages in the base library already have the version 0.1,
but they differ from the output currently generated by the script.
So we need to bump the version before we update these packages in the
base library.

Also fixed the creation date which currently does not correspond to the
date of the packages already added to the base library. So this fix
ensures that re-generating these packages does not change their
creation date.