Generate output into corresponding library directory (#103)
Instead of generating the output into a directory like "out/led/pkg/",
use the actual library directory where the files need to be added to,
for example "out/LibrePCB_Base.lplib/pkg". This makes it much easier
to copy the generated elements into the libraries. In addition, it
is less error-prone since the library assignment is now stored in the
generators (currently it's cumbersome to find out which files need to
be copied into which library).